requirements.txt
streamlit==1.29.0
yfinance==0.2.33
python-dotenv==1.0.0
plotly==5.18.0
pandas==2.1.4
requests==2.31.0
numpy>=1.24.3
